feat: Added helper utilities
The new util files add dummy data to db with photos correctly. It also has an ipynb file that can load an image from db to verify.

Also added indices for faster searches and an experimental schemas. DO NOT use inserts.sql to populate the db. It does not salt and hash the password